# apps/server/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/articles":{"get":{"operationId":"ArticleController_findAll","summary":"Get all articles","parameters":[],"responses":{"200":{"description":"Return all articles."}},"tags":["articles"],"security":[{"bearer":[]}]},"post":{"operationId":"ArticleController_create","summary":"Create article","parameters":[],"responses":{"201":{"description":"The article has been successfully created."},"403":{"description":"Forbidden."}},"tags":["articles"],"security":[{"bearer":[]}]}},"/api/articles/feed":{"get":{"operationId":"ArticleController_getFeed","summary":"Get article feed","parameters":[],"responses":{"200":{"description":"Return article feed."},"403":{"description":"Forbidden."}},"tags":["articles"],"security":[{"bearer":[]}]}},"/api/articles/{slug}":{"get":{"operationId":"ArticleController_findOne","parameters":[],"responses":{"200":{"description":""}},"tags":["articles"],"security":[{"bearer":[]}]},"put":{"operationId":"ArticleController_update","summary":"Update article","parameters":[],"responses":{"201":{"description":"The article has been successfully updated."},"403":{"description":"Forbidden."}},"tags":["articles"],"security":[{"bearer":[]}]},"delete":{"operationId":"ArticleController_delete","summary":"Delete article","parameters":[],"responses":{"201":{"description":"The article has been successfully deleted."},"403":{"description":"Forbidden."}},"tags":["articles"],"security":[{"bearer":[]}]}},"/api/articles/{slug}/comments":{"get":{"operationId":"ArticleController_findComments","parameters":[],"responses":{"200":{"description":""}},"tags":["articles"],"security":[{"bearer":[]}]},"post":{"operationId":"ArticleController_createComment","summary":"Create comment","parameters":[],"responses":{"201":{"description":"The comment has been successfully created."},"403":{"description":"Forbidden."}},"tags":["articles"],"security":[{"bearer":[]}]}},"/api/articles/{slug}/comments/{id}":{"delete":{"operationId":"ArticleController_deleteComment","summary":"Delete comment","parameters":[],"responses":{"201":{"description":"The article has been successfully deleted."},"403":{"description":"Forbidden."}},"tags":["articles"],"security":[{"bearer":[]}]}},"/api/articles/{slug}/favorite":{"post":{"operationId":"ArticleController_favorite","summary":"Favorite article","parameters":[],"responses":{"201":{"description":"The article has been successfully favorited."},"403":{"description":"Forbidden."}},"tags":["articles"],"security":[{"bearer":[]}]},"delete":{"operationId":"ArticleController_unFavorite","summary":"Unfavorite article","parameters":[],"responses":{"201":{"description":"The article has been successfully unfavorited."},"403":{"description":"Forbidden."}},"tags":["articles"],"security":[{"bearer":[]}]}},"/api/user":{"get":{"operationId":"UserController_findMe","parameters":[],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]},"put":{"operationId":"UserController_update","parameters":[],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/api/users":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/api/users/{slug}":{"delete":{"operationId":"UserController_delete","parameters":[],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/api/users/login":{"post":{"operationId":"UserController_login","parameters":[],"responses":{"201":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/api/profiles/{username}":{"get":{"operationId":"ProfileController_getProfile","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["profiles"],"security":[{"bearer":[]}]}},"/api/profiles/{username}/follow":{"post":{"operationId":"ProfileController_follow","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["profiles"],"security":[{"bearer":[]}]},"delete":{"operationId":"ProfileController_unFollow","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["profiles"],"security":[{"bearer":[]}]}},"/api/tags":{"get":{"operationId":"TagController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["tags"],"security":[{"bearer":[]}]}},"/api/upload/album":{"post":{"operationId":"UploadController_upload","summary":"上传文件","parameters":[],"responses":{"200":{"description":"上传成功"}},"tags":["upload"],"security":[{"bearer":[]}]}},"/api/upload/export/{filename}":{"get":{"operationId":"UploadController_downLoad","summary":"下载文件","parameters":[],"responses":{"200":{"description":"下载成功"}},"tags":["upload"],"security":[{"bearer":[]}]}}},"info":{"title":"NestJS Realworld Example App","description":"The Realworld API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{}}}}}